// src/bridge/generated/bindings.ts
export type EditorMode = "translation" | "proofreading" | "readonly";
// ...
export type SpecialCharacter = {
	id: string,
	text: string,
	is_favorite: boolean,
};

export type TextStage = "translation" | "proofreading";

export type Unit = {
	id: string,
	page_id: string,
	index: number,
	x_coord: number,
	y_coord: number,
	is_bubble: boolean,
	is_flagged: boolean,
	translated_text: string,
	proofread_text: string,
	is_proofread: boolean,
	created_at: number,
	updated_at: number,
};

export type UnitDraft = {
	id: string,
	x_coord: number,
	y_coord: number,
	is_bubble: boolean,
	is_flagged: boolean,
	translated_text: string,
	proofread_text: string,
	is_proofread: boolean,
};